elf: Move ELF types to separate module
In the effort of splitting up the ELF parser code, move the ELF types
to their own module.
