+++ elf/src/types.rs
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: (GPL-2.0-or-later OR MIT)
+//
+// Copyright (c) 2023 SUSE LLC
+//
+// Author: Nicolai Stange <nstange@suse.de>
+//
+// vim: ts=4 sw=4 et
+
+pub type Elf64Addr = u64;
+pub type Elf64Off = u64;
+pub type Elf64Half = u16;
+pub type Elf64Word = u32;
+#[allow(unused)]
+pub type Elf64Sword = i32;
+pub type Elf64Xword = u64;
+pub type Elf64Sxword = i64;
+pub type Elf64char = u8;
